refactor: create helpers/books.ts module and replace here functions connected with books
